myproject/app1/serializers.py
from rest_framework import serializers
from .models import Product, Order

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'  # หรือระบุ fields ที่ต้องการ เช่น ['id', 'name', 'price']


class OrderStatusSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = ['user', 'status', 'total_price']