requirements.txt
Django==4.2
djangorestframework>=3.15.2


psycopg2-binary
pillow
python-dotenv
social-auth-app-django
segno
crcmod>=1.7
promptpay
qrcode